chore(SingleEventView): renames component to keep names consistent
